Revert "Merge https://github.com/MIPT-ILab/mipt-mips into Compressed-ISA"
This reverts commit 78e22a8c9adc2ff95ae6d1d809d03f0b95a5f56c, reversing
changes made to 49b577256411d064a8ff2b206437d7414ddb6958.
