+++ simulator/infra/cache/cache_tag_array.cpp
@@ -79,15 +79,9 @@ CacheTagArray::CacheTagArray(
     uint32 addr_size_in_bits)
     : CacheTagArraySize( size_in_bytes, ways, line_size, addr_size_in_bits)
     , tags( sets, std::vector<Tag>( ways))
-    , lookup_helper( sets, google::dense_hash_map<Addr, uint32>( ways))
+    , lookup_helper( sets, std::unordered_map<Addr, uint32>( ways))
     , replacement_module( sets, ways)
-{
-    //theese are spicial dense_hash_map requirements
-    for (uint32 i = 0; i < sets; i++) {
-        lookup_helper[i].set_empty_key( impossible_key);
-        lookup_helper[i].set_deleted_key( impossible_key - 1);
-    }
-}
+{ }
 
 std::pair<bool, uint32> CacheTagArray::read( Addr addr)
 {
@@ -143,3 +137,5 @@ ReplacementModule::ReplacementModule( std::size_t number_of_sets, std::size_t nu
     for ( std::size_t i = 0; i < number_of_sets; i++)
         replacement_info.emplace_back( create_cache_replacement( replacement_policy, number_of_ways));
 }
+
+
