+++ simulator/infra/types.h
@@ -47,16 +47,9 @@ using uint128 = unsigned __int128;
 
 static inline std::ostream& operator<<(std::ostream& out, uint128 value)
 {
-    static const uint128 separator = 10'000'000'000'000'000'000u;
-
-     if ((out.flags() & std::ios::hex) != 0)
+    if ((out.flags() & std::ios::hex) != 0)
         return out << narrow_cast<uint64>( value >> 64ULL) << narrow_cast<uint64>( value);
-
-    if (value <= UINT64_MAX)
-        return out << narrow_cast<uint64>( value);
-    uint128 leading  = value / separator;
-    uint64  trailing = value % separator;
-    return out << leading << trailing;
+    return out << "..." << narrow_cast<uint64>( value);
 }
 
 #pragma GCC diagnostic pop
